Extend WebLayer API to support content filtering. (#2)
Enables consumption of easylist style rules encoded using components/url_pattern_index as a protobuf.

ContentFilteringService is an interface provided by the browser to renderer proceses, which use it to fetch the protobuf via a file descriptor that can be memory mapped.

This enables the protobuf to be stored as part of the application APK and efficiently shared to renderer processes without any copies. (Important for the easylist protobuf files to not be compressed as part of the APK however.)

ContentFilterManager is the main controller interface used by the embedding app to configure the content filter. The embedding app can also observe when filtering happens by attaching a ContentFilterCallback instance to each Tab.

There's a command line tool called make_filter_rules that takes an easylist format file and outputs the corresponding protobuf.