Add support for packaging WebLayerSupport as a feature module (#4)
Changes WebLayerImpl to handle being loaded either as part of WebLayerSupport.apk or as an APK split named weblayer_support. WebLayerImpl tells the difference by checking to see what package the given appContext and removeContext belong to.

WebLayer client code is updated accordingly and importantly uses DelegateLastClassLoader (instead of PathClassLoader) to ensure that the dexPath of the split takes precedence over the dexPath of the base ClassLoader. ART gets unhappy if the parent ClassLoader does not correspond to the base APK split, so this is how we navigate that constraint.

It is critical that classes from the weblayer_support APK split get preferred by WebLayer code so that the splits can safely include duplicate symbols -- different versions of AndroidX libraries, etc.

It is also necessary for the easyprivacy list to be included as a resource of WebLayerSupport so that it can be looked up using ContextUtils.getApplicationContext(), so it is included in this PR. Had to switch from .proto suffix to .dat since the former is a text file format (protobuf definition), and Chromium GIT is configured to munge line endings for text files.

This change breaks support for using the system WebView in the same process that uses WebLayer.