Use proper package name to resolve resources (#6)
For icons in notifications, WebLayer uses the Icon.createWithResource
method that takes a package name and a resource ID. (Not sure why it
doesn't use the createWithResource variant that takes a Context.)

When WebLayer is bundled as a feature module, the package name ends up
being that of the embedding application.

The resource IDs that ordinarily have a 0x24 prefix need to be adjusted
so that the system can resolve those IDs against the given package name
of the embedding app. The forceAddAssetPaths method, used by WebLayer
to make resources with a prefix of 0x24 work, does not work in this case
probably because another app is trying to resolve the resource IDs using
a different context.

That means we need to know which package ID is assigned to the WebLayer
APK. In both configurations, both as a feature split and when we are
just using WebLayerSupport.apk directly, that can be done by querying
the package ID of "org.chromium.weblayer.support".

Then we can adjust the resource IDs to that package ID and we are good.