Rework how rules are passed to renderers and support null
Previously the browser process would happily send back a null rules object
to the renderer, but because RefreshRules was declared to not receive a
null rules object, the Mojo connection would be disconnected. As a result,
the renderer would never be told to stop filtering (which is what a null
rules object means).

Issue neevaco/neeva-android/issues/629