Rework how rules are passed to renderers and support null (#7)
Previously the browser process would happily send back a null rules object
to the renderer, but because RefreshRules was declared to not receive a
null rules object, the Mojo connection would be disconnected. As a result,
the renderer would never be told to stop filtering (which is what a null
rules object means).

This PR also simplifies how new rules are passed to a renderer. Instead of
the old clunky callback system with generation numbers, now a listener
interface is implemented by each ContentFilteringAgent and passed to
the service via a new AddRulesListener method. (This is not essential to
fixing the bug.)

Also while updating the IPC, removes the unnecessary Log method.
Normal base/logging works just fine from the renderer process.

Issue neevaco/neeva-android/issues/629