Ad blocker support (#11)
Adds support in the API for specifying multiple rules files, so we can enable
both easylist and easyprivacy. In the future we might want to expand to include
other lists too.

Adds support for filtering out DOM elements from given CSS selectors, which
is needed to support easylist ad blocking. This is done by generating a user
stylesheet that is a concatenation of selectors that get the same display:none
rule applied to them.

Adds a small renderer-side cache of generated stylesheets to cut down on the
cost of constructing these stylesheets. Uses flat buffers to serialize the CSS
rules into a format that can be somewhat efficiently consumed in the renderer
processes. This is inspired from how url_pattern_index uses flat buffers.