Allow clients to query filter stats by rules file
Extend the WebLayer API to report content filtering stats on a per filter list basis.
This will enable the embedding app to determine if any ads were blocked or if
only trackers were blocked, for example.

The approach of providing an API to query by passing the name of the filter list
helps keep the implementation simple. An alternative would be to return the entire
data set in one go, but bridging that across the AIDL divide is a bit complicated.